pyproject.toml
[tool.poetry]
name = "youtube2podcast"
version = "0.1.0"
description = ""
authors = ["Artur"]

[tool.poetry.dependencies]
python = "^3.9"
pytube = "^12.0.0"
kafka-python = "2.0.2"
python-telegram-bot = "13.11"
moviepy = "1.0.3"
pydantic = "1.9.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

